requirements_base.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==2.0.6
django==5.2.9
django-celery-beat==2.8.1
django-extensions==4.1
django-formtools==2.5.1
django-timezone-field==7.2.1
et-xmlfile==2.0.0
idna==3.11
kombu==5.6.2
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
xlrd==2.0.2
